approval-gateway/requirements.txt
# Core dependencies
fastapi==0.115.5
uvicorn[standard]==0.34.0
pydantic==2.10.3
pydantic-settings==2.6.1

# Redis
redis==5.2.0

# HTTP client
httpx==0.28.1

# Twilio for WhatsApp
twilio==9.3.7

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-mock==3.14.0

# Linting & formatting
ruff==0.8.4
black==24.10.0
mypy==1.13.0

# Utilities
python-dotenv==1.0.1

